Add/Sub/Mul/Div for all builtin CoordinateTuple types
