+++ src/coordinate/coor4d.rs
@@ -1,5 +1,4 @@
 use crate::coordinate::*;
-use std::ops::{Add, Div, Mul, Sub};
 
 /// Generic 4D coordinate tuple, with no fixed interpretation of the elements
 #[derive(Debug, Default, PartialEq, Copy, Clone)]
@@ -20,68 +19,6 @@ impl CoordinateTuple for Coor4D {
 
     fn set_nth_unchecked(&mut self, n: usize, value: f64) {
         self.0[n] = value;
-    }
-}
-
-// ----- O P E R A T O R   T R A I T S -------------------------------------------------
-
-impl Add for Coor4D {
-    type Output = Self;
-    fn add(self, other: Self) -> Self {
-        Coor4D([
-            self.0[0] + other.0[0],
-            self.0[1] + other.0[1],
-            self.0[2] + other.0[2],
-            self.0[3] + other.0[3],
-        ])
-    }
-}
-
-impl Add<&Coor4D> for Coor4D {
-    type Output = Self;
-    fn add(self, other: &Self) -> Self {
-        Coor4D([
-            self.0[0] + other.0[0],
-            self.0[1] + other.0[1],
-            self.0[2] + other.0[2],
-            self.0[3] + other.0[3],
-        ])
-    }
-}
-
-impl Sub for Coor4D {
-    type Output = Self;
-    fn sub(self, other: Self) -> Self {
-        Coor4D([
-            self.0[0] - other.0[0],
-            self.0[1] - other.0[1],
-            self.0[2] - other.0[2],
-            self.0[3] - other.0[3],
-        ])
-    }
-}
-
-impl Mul for Coor4D {
-    type Output = Self;
-    fn mul(self, other: Self) -> Self {
-        Coor4D([
-            self.0[0] * other.0[0],
-            self.0[1] * other.0[1],
-            self.0[2] * other.0[2],
-            self.0[3] * other.0[3],
-        ])
-    }
-}
-
-impl Div for Coor4D {
-    type Output = Self;
-    fn div(self, other: Self) -> Self {
-        Coor4D([
-            self.0[0] / other.0[0],
-            self.0[1] / other.0[1],
-            self.0[2] / other.0[2],
-            self.0[3] / other.0[3],
-        ])
     }
 }
 
@@ -161,6 +98,7 @@ impl Coor4D {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::ops::{Add, Div, Mul};
 
     #[test]
     fn distances() {
